Warning in small views

+++ app/src/main/java/cz/lhoracek/android/dials/utils/BindingAdapters.java
@@ -19,7 +19,7 @@ public class BindingAdapters {
     }
 
     @BindingAdapter("warningMax")
-    public static void setWarningMax(ValueView valueView, int warning){
+    public static void setWarningMax(ValueView valueView, float warning){
         valueView.setWarningMaxValue(warning);
     }
